geoip: update the GeoIP databases on mirrorbits reload (close #30)
On mirrorbits reload (SIGHUP), the modification dates of the on-disk GeoIP
databases will be compared and if different, will be reloaded. In case
of failure, the previous databases will be kept in memory.

The reload can be automated by editing the geoip-lite-update script
available in the contrib directory.

Reloading the 4 databases takes around 15-20ms on my test machine and will
delay some of the in-flight requests by this amount.