Keep track of the sync offset between mirrors (closes #1)
This new feature will compute the time offset between the main
repository and all the mirrors allowing to see in a glance which
mirrors are up-to-date and which are not (in a similar way to mirmon).

Like mirmon, it requires the master repository to contain a file that
contains a unix timestamp updated at regular interval. For example, to
update that file every five minutes:
  */5 * * * * date +%s > /srv/repository/trace

The file will be mirrored by all the mirrors and mirrorbits will use it
to compare the offset since their last rsync synchronization. The more
often the file is updated more the results will be accurate. Mirrorbits
will check the trace on each mirror using HTTP therefore don't use hidden
files since some mirror don't allow hidden files to be accessed.

The frequency of the updates from mirrorbits is based around the
ScanInterval value. The file name and location can be changed using the
TraceFileLocation value and the result is visible on the mirrorstats page.