service: use systemd DynamicUser isolation
DynamicUser allows systemd to create a temporary user to manage the
lifetime of the service. Therefore it is not necessary to create a new
mirrorbits user in order to drop privileges. Most of the filesystem is
also read-only and some system folders like /tmp are isolated from other
applications. The only persistent storage contains the logs, their
permissions are managed automatically by systemd (if the uid ever change)
but everything else is discared when the process is gone.

This is now the most secure way to run Mirrorbits in production.

See http://0pointer.net/blog/dynamic-users-with-systemd.html