service: Support systemd notify type
With the current default of Type=simple, systemd considers mirrorbits
started as soon as the mirrorbits process starts. However, neither the
RPC nor HTTP ports are bound at that point, so services that start After
mirrorbits may fail if systemd starts them immediately.

Change to Type=notify and send the READY=1 message through the notify
socket after the ports are bound but just before the HTTP server starts.
If systemd is not in use (the NOTIFY_SOCKET environment variable is not
set), then this is a no-op.