http: run "SET" with the "GET" option for the range query check.
This option only appeared in redis 6.2.0. Note that there is the
deprecated GETSET command, and we could have used it to cater to all
redis versions, except that the EX option doesn't exist on GETSET. So it
won't do.

Instead let's just run the single SET call on redis 6.2.0 and over, and
continue to run 2 requests on older redis databases.

This uses the fact we now store the redis version (cf. previous commit),
otherwise if we were to query the version each time (+ 1 request), it
would make no sense.