Add a contrib script to delete old mirrorbits stats
Otherwise stats pile up forever, stored in the Redis database, hence
eating precious RAM.

In practice, the keys MIRROR_${date} and MIRROR_BYTES_${date} don't take
too much space. Each key is a hash, and the number of items in the hash
is the number of mirrors, so that shouldn't be too big.

However, the keys FILE_${date} are more problematic. Each key is a hash,
but this time the number of items in the hash is the number of files
that were served. For big repositories (think Linux distribution for
which the repo has between 500k to 1 million files), each daily stat can
be quite big, so if we keep it all in the database, in the long-term it
will really waste a considerable amount of RAM.

It would be nice if Mirrorbits would provide a command to cleanup the
old stats. In the meantime, this script should do the job just as well.