Add RelaxModTimeRules config option
With this setting, files are allowed to be outdated on the mirrors, for
at most MaxOutdated minutes. The filesize check is also disabled for
those files.

Use-case: for a Debian-like distribution, the metadata (ie.  the
directory /dists) are updated in-place, so we must give time for mirrors
to sync, and then for mirrorbits to be aware of the changes.

Otherwise, as soon as the source is updated and scanned, Mirrorbits will
go in fallback mode for all the files under /dists, since at this point,
either mirrors didn't sync yet, either they did but Mirrorbits is not
aware of it yet (as the interval to scan mirrors is higher than the
interval to scan the source).

Cf. https://github.com/etix/mirrorbits/issues/85 for more details.