Dual: commands: list/stats: Update status display
Now that a mirror can have two URLs (HTTP and HTTPS), it's a bit less
straightforward to decide if the mirror is up or down, and to display
the state.

A mirror is said to be "up" if:
- it has two URLs, and both are up
- it has only one URL, and it's up

In order to display the state of the mirror: we still use up and down if
ever the mirror is fully up or fully down. But in case it's up for one
URL, and down on the other, we introduce the new state:

    <http-state>/<https-state>

Hence a mirror might be reported as:

    up/down    <- up over HTTP, down over HTTPS
    down/up    <- down over HTTP, up over HTTPS