Dual: commands: add: Add support for https
Add a new flag -https to specify the HTTPS URL.

When adding a mirror, either -http or -https (or both) must be passed,
otherwise mirrorbits fails.

Mirrorbits now checks for protocol mismatch, and fails if you pass a
HTTPS URL to -http, or a HTTP URL ot -https.

Finally, mirrorbits makes sure that the two URLs point to the same host.
This is because we rely on the mirror hostname to geolocate it, so if we
had two URLs pointing to two different hosts, it wouldn't work.