Dual: rpc: Pick the right URL to geolocate a mirror
We pick a non-empty URL, favoring the HTTP URL if both are set, in order
to get the host name and then geolocate the mirror.

It makes the assumption that both HTTP and HTTPS URLs point to the same
host, which was addressed in previous commits (we check that in the add
and edit commands).