Dual: mirrors, monitor: Add a IsUp method to mirror, use it in monitor
Add a IsUp() method to mirrors. This is the same as the IsUp() function
that was added a few commits ago in cli/commands.go, with the difference
that in cli/commands.go, IsUp() acts on a rpc.Mirror type, while here,
IsUp() is a method of mirrors.Mirror.

We use this function in daemon/monitor.go, to enqueue a health-check if
a mirror is not up.