Dual: mirrors, monitor, logs: Rework the health-check
Now that a mirror can have two URLs (HTTP and HTTPS), the health-check
needs to be reworked, and basically duplicated.

In monitor, we move most of the healthCheck() function in another
function healthCheckDo(). Very little remains in healthCheck(): it only
picks a random file, and then it calls healthCheckDo() for the HTTP URL
(if set), and then again for the HTTPS URL (if set).

It's important to use the same file for both health checks, so that if
ever one check fails and the other succeeds, we don't ask ourselves if
it's because of the file that was picked.

Then, doing both checks sequentially is not super elegant, but it's also
super easy to implement it this way, without having to rewrite too much
code, so let's live with that.

Now for an overview of the changes in healthCheckDo():
- we now include the protocol in the log output
- we need to pass the protocol to the functions MarkMirror{Up,Down}

Which leads us to the changes in mirrors/mirrors.go. We introduce a
Protocol enum, and we use it for MarkMirror{Up,Down} and SetMirrorState.

The function SetMirrorState becomes a little bit more tricky, and we
take this chance to check that the URL is set when user tries to set the
state to 'up'. This is because, obviously, if a mirror has no URL for a
protocol (HTTP or HTTPS), then the only legit state for this protocol is
'down'.

Finally, we also update mirrors/logs.go to include the protocol in the
LogStateChanged logs. Going forward, the lines:

    Mirror is up/down ...

will be replaced with

    HTTP mirror is up/down ...

and

    HTTPS mirror is up/down ...

For old logs already in the database, the protocol is not stored, and
that will be translated to a value of 0 aka. UNDEFINED, which will be
logged as 'Mirror is up/down ...' (that is: as before).