Dual: Update mirrorlist.html template
We make use of the variable mirrors.Mirror.SelectedProtocol that was
introduced in the previous commit. This variable indicates which
protocol was selected for redirection, for a given mirror.

The logic to get the right URL in the template is a bit tricky, it's a
best effort to pick a non-empty URL, and preferably the right one.