Dual: Add a database upgrade (v1 -> v2)
This upgrade only changes the mirrors, ie. the MIRRORS key in the redis
database. It does the following:

- always: add new field `https`
- optionally: add new field `httpsUp` (if HTTPS URL and `up` exists)
- optionally: add new field `httpsDownReason` (if HTTPS URL and
  `excludeReason` exists)
- optionally: rename `up` -> `httpUp` (if HTTP URL and `up` exists)
- optionally: rename `excludeReason` -> `httpDownReason` (if HTTP URL
  and `excludeReason` exists)

In case it's not clear: the keys `up` and `excludeReason` are optional,
they might not exists in the database (in case a mirror was never
scanned or enabled). Hence we create their counterparts `https?Up` and
`https?DownReason` only if needed.