Dual: http: Make fallbacks respect the protocol
As a reminder, mirrorbits behaviour is as such:
- HTTP request : mirrorbits redirects to either HTTP or HTTPS, depending
  on what's available
- HTTPS request: mirrorbits redirects to HTTPS

Before this commit, the fallback URLs were used "as is". This was
actually not perfect, for example if your fallback mirror was defined as
http://foobar.org, then mirrorbits would happily redirect HTTPS requests
to a HTTP URL.

Of course, the right thing to do is to specify only HTTPS fallback
mirrors, but it might not be obvious for everyone, especially since the
sample mirrorbits.conf doesn't mention that, and gives a HTTP fallback
mirrors (http://fallback1.mirror/repo/) as an example.

With this commit, we change that a bit.

First, in the sample mirrorbits.conf, we mention that fallback mirrors
should support both HTTP and HTTPS (which, I think, is a reasonable
requirement).

Then, in the code, we now make the fallback selection a bit smarter,
possibly ignoring the protocol that was set by user in the conf:
- if WITHTLS: force HTTPS
- if WITHOUTLS: force HTTP
- otherwise: use the URL "as is"

This logic for the protocol selection is what's used already when
selecting mirrors, so basically this commit brings fallback mirrors in
lines with mirror selection.