Prevent yaml.Marshal from wrapping long lines
Line wrapping after 80 characters was the default in yaml.v2. No line
wrapping is the default for yaml.v3. The helper FutureLineWrap() is
meant to help users do the transition to the new behavior.

For mirrorbits, line wrapping manifests itselfs for the `show` and
`edit` commands. It's an issue, in the sense that it makes it difficult
for scripts to interact with those commands (as long values might spread
on several lines, so a simple grep won't do).

Hence this commit disables line wrapping.

Closes: #153