Dual: selection: Assign the mirror copy (rather than orig) in mlist
This is a subtle change of no consequence in the current version of the
code, but that is needed for the next commit.

The loop here use copies, as `mslist` is a slice of mirrors (and not of
mirror pointers), hence `for i, m := range mlist` creates copies. In
other words, `m` is a copy of `mlist[i]`.

That brings a question: when we save the mirror in one of the result
slices (`mlist` for a mirror that is selected, and `excluded` for a
mirror that is excluded), what should we store in the slice? The
original `mlist[i]`, or the (potentially modified) copy `m`?

As we can see, in the `excluded` slice, we clearly use `m`, in the line:

    excluded = append(excluded, m)

This is needed, as `m` is modified to include the ExcludeReason, for
example in `m.ExcludeReason = "Invalid URL"`.

However, when assigning to `mlist`, we use the original (unmodified)
mirror:

    mlist[safeIndex] = mlist[i]

Is there a good reason to use the original, rather than the copy? After
a careful read of the code, we can see that it doesn't matter: when a
mirror is selected (hence added to `mlist`), it has not been modified.
So, at this point, `mlist[safeIndex] = mlist[i]` is equivalent to
`mlist[safeIndex] = m`.

So why change it? This in preparation for the next commit, where we'll
introduce the field `m.SelectedProtocol`. This new field will be set for
mirrors that are selected. So, starting next commit, we'll modify the
copy of the mirror in both cases (selected or excluded), so we must save
the copy (rather than the orig) in both cases as well.

We do the change now, as it's easier to review.