Dual: mirrors, pagerenderer, selection: Rework mirror selection
In this commit, we introduce a SelectedProtocol field in the mirror
struct.  The best way to look at it is to look at mirrors/mirrors.go and
at http/pagerenderer.go.

It is needed for the case where neither HTTP nor HTTPS is forced (ie.
the incoming request was over HTTP, so mirrorbits will happily redirect
to either HTTP or HTTPS, depending on what's available). If ever a
mirror provides a HttpURL AND a HttpsURL, we need to remember which URL
was selected, so that the code in http/pagerenderer.go is able to
correctly build the redirections.

Then the second part of this commit is the refactoring in
http/selection.go, to take into account the fact that a mirror can now
have a HTTP and/or a HTTPS URL. This is rather dumb code, easy to read.