Switch to gopkg.in/yaml.v3
The main reason to switch (apart from keeping dependencies up-to-date)
is the default behaviour for line wrapping. In yaml.v2, the default was
line wrapping after 80 characters. In yaml.v3, no line wrapping by
default.

For mirrorbits, line wrapping manifests itself in the `show` and `edit`
commands. It's an issue, in the sense that it makes it difficult for
scripts to interact with those commands, since long values might spread
over several lines. For example, parsing the output of `mirrorbits show`
with `grep` is error prone because of line wrapping.

This commit bumps the build depencency gopkg.in/yaml from v2 to v3, thus
disabling line wrapping in yaml output.

Closes: #153