Add -h / -help flag for commands reload, upgrade and version
This commit adds support for the `-h / -help` flag for the 3 commands
that didn't have it: `reload`, `upgrade` and `version`.

It looks like an unsignificant change, but it's a bit more than that.

Most commands accept -h (or -help), so typing `mirrorbits <command> -h`
becomes an automatism. Except that when one wants to learn a bit more
about the `upgrade` command, surprise! `mirrorbits upgrade -h` actually
executes the upgrade command. No message printed on the console, as this
command is completely silent.

I'm not sure exactly what is the expected outcome of running `mirrorbits
upgrade`, but in my case, mirrorbits quited and didn't restart. So I
just took the production instance down while trying to display a help
message, and I didn't even notice.

I think it's much safer if the `upgrade` command supports `-h / -help`,
and for consistency, let's also add it for the `reload` and `version`
commands.