http: Sync If-Modified-Since with latest go/src/net/http/fs.go
The code for If-Modified-Since was proposed a while ago, and since then
there were two changes upstream, so let's sync.

1. Include change 59ab6f351a370a27458755dc69f4a837e55a05a6:

```
+	delete(h, "Content-Encoding")
```

2. Do NOT include change 879f595f7eacbd53d25fe21cac4b2b0cfde36449:

```
-	if modtime.Before(t) || modtime.Equal(t) {
+	if ret := modtime.Compare(t); ret <= 0 {
```

`func (Time) Compare` was introduced in Go 1.20 released in Feb 2023,
that's rather new. In Debian stable (bookworm) we have Go 1.19. So let's
not include this commit, and let people build mirrorbits with older
versions of Go if they need to.