cache_test: Fix the last test in TestCache_GetFileInfo
On the second call to GetFileInfo, we want to ensure that we hit the LRU
cache. The best way to do that is to ensure that the HMGET Redis command
is NOT called a second time. Which is easy to do, but was not done, so
we do it with this commit.

And at the same time, we also drop the check for Redis.ErrNil. This
check doesn't make much sense actually, since it comes after the second
call to GetFileInfo, meaning the function was called successfully once
already. Why would we get ErrNil on the second call?

Plus, as you'll see a few commits after, GetFileInfo never returns
Redis.ErrNil anyway...