http: Drop useless check for Redis.ErrNil
This is a follow-up of previous commit, where we added new unit tests
that show that GetFileInfo doesn't return Redis.ErrNil.

Therefore we can drop this useless check.

Grepping through the code, there are a few places where GetFileInfo is
called, but it's only in this one place that there was a check for
Redis.ErrNil.

Just to clarify, in case it's not clear: this commit doesn't change
mirrorbits' behavior, it really removes a check that never triggered.
After this commit, querying the checksum for a non-existing file still
returns 404 "Not Found", due to the check `if len(hash) == 0` that comes
a few lines below.