http: Get file information earlier, don't wait for selection
This is in preparation for the next commit, where we'll add support for
If-Modified-Since. For that, we need to know the file modTime, so we
need to call GetFileInfo() early enough.

Calling the function earlier doesn't hurt, and doesn't change anything
to the logic of the code. As we can see, before this commit:
- if the function fails, we bail out of Selection() an error, it take us
  to the line `http.Error(w, err.Error(), http.StatusInternalServerError)`
- if the function succeeds, we proceed. That includes the case where a
  file is not found in the Redis DB.

With this commit, this logic is preserved.

Extra notes: (1) `filesystem.NewFileInfo(urlPath)` returns a struct
where _only_ Path is populated, while (2) `cache.GetFileInfo(urlPath)`
returns a struct where _at least_ Path is populated. So calling (1) then
(2) is the same as just calling (2).