http: Set Last-Modified header when returning 304 "Not Modified"
A client that sends If-Modified-Since probably cares about the
Last-Modified time, so let's add it in the response header.

Actually, I have the impression that we could set this header all the
time, and not only when the request includes If-Modified-Since (assuming
the file requested was found in the redis db). However I have no idea
what's the pros and cons, all I can say is: we have the information, so
why not share it with the client?

Maybe some HTTP guru have an opinion here.

Note that this code was also picked "as is" from go/src/net/http/fs.go.