cache_test: Add a helper to ease comparing filesytem.FileInfo
With this helper, it's now easy to compare the actual struct with the
expected struct. So we can do a proper check everywhere needs be.

In particular we test that the value returned by the LRU cache is
correct, which wasn't done before.

This is also in preparation for next commit.