New settings SameDownloadInterval.
This lets admins customize the time interval (in seconds) between which
they want to consider 2 range downloads to be the same download.

Setting SameDownloadInterval to 0 disables the same download detection
logic (i.e. that every range download counts as a full download, just as
it used to happen until now).