cli: Try harder to find a suitable text editor
Currently the command `mirrorbits edit` errors out when the EDITOR
environment variable is not set:

```
$ mirrorbits edit mirror.example.com
2024/12/17 14:07:58 Environment variable $EDITOR not set
```

It's not difficult to do better, we can check for the most common text
editors. `editor` is checked first, it should be a symlink to the
prefered editor on Debian systems (configured by user).

Note that this code was not invented here, it's taken verbatim from
incus source code, with thanks.