monitor: Move healthCheck implementation in another function
With this change, the actual health check implementation (ie. the http
request) is moved to a new function called `healthCheckDo`. The original
`HealthCheck` function does only two things now: 1) pick the file that
will be used to perform the health check and 2) call `healthCheckDo`.

The purpose of this split is to allow, in the upcoming commits, to
perform 2 health checks instead of one. That will be used for mirrors
that support both HTTP and HTTPS, we will need to check that requests
work for both protocols.