selection: Move filtering in a separate Filter function, add tests
In this commit, all the code block that filter mirrors (ie. separate
them in two groups, mirrors candidates for the redirect, and mirrors
excluded) is moved to a separate function, for the purpose of adding
test cases for it.

The test cases added don't cover every possible situation, far from
that, but at least it's a start.