Add trivial helpers to know if a mirror is up, or to give its status
Those functions are trivial for now, but in the upcoming commits we will
allow a mirror to support both HTTP and HTTPS. The logic to know if it's
up will be more complex, given that a mirror can be up over HTTP but
down over HTTPS, for example.

Similarily, a status string will be a bit more than just "up" or "down",
eg. when the mirror is up for HTTP and down for HTTPS.