mirrorbits.conf: Set the fallback protocol to https
This change is just to nudge the user in the right direction.

The behavior of mirrorbits, with regards to protocol, is as such:
* HTTPS requests: redir to HTTPS mirrors only
* HTTP  requests: redir to HTTP and HTTPS mirrors

When it comes to redirecting to the fallback mirror however, mirrorbits
doesn't care about the protocol of the incoming request, it redirects
without any check. It means that, if fallback mirrors are HTTP,
mirrorbits will happily redirect HTTPS to HTTP, which is of course very
wrong.

In other words, from the moment a mirrorbits instance supports HTTPS,
the fallback mirror must have a HTTPS URL as well. Most users probably
want HTTPS fallback mirrors, so let's show that in the default
mirrorbits.conf.