Add AbsoluteURL field in Mirror
This is in preparation to support relative URLs.

This commit adds a new field `AbsoluteURL` in the `Mirror` struct
(`mirrors/mirrors.go`).

This field is derived from the field `HttpURL`, and is guaranteed to be
a absolute URL (ie. to have a scheme).  At the moment, it's equal to
`HttpURL` (since `HttpURL` always have a scheme), but that will change
in the next commits.

As can be seen (or rather, guessed) in the Mirror struct, this field is
among the fields that are set during mirror selection, so only the code
that comes after that can make use of it, and it is not saved in the
redis db.

In practice, this field is set in two places: `http/selection.go` and
also in `http/http.go` (in the code that handles fallbacks mirrors).