Add support for HTTP+HTTPS mirrors
In previous commit "Support relative URL in the mirror HttpURL" we
started to accept mirror URLs without a scheme, eg. mirror.com/foobar/,
but in this case the scheme was HTTP by default, so it was not very
interesting...

With this commit, we change this behaviour: mirrors with a relative URL
now support both HTTP and HTTPS. It means that the health check is done
twice (for HTTP and for HTTPS, cf. `daemon/monitor.go`). We duplicate
the fields `Up, DownReason` to become `HttpUp, HttpsUp, HttpDownReason,
HttpsDownReason` in the Mirror struct `mirrors/mirrors.go`). And finally
rework the selection logic to redirect appropriately
(`http/selection.go`).

Note that, given that some redis keys are renamed, there needs to be a
database migration - that will come in a later commit.