mirrorbits-del-stats: Update usage message
Add missing `--` in example.

Use `-p xxx` in the example, instead of `-n xxx`. The example wants to
show how to deal with multiple mirrorbits instance, and in that case
it's better to have multiple redis servers (so `-p` to talk to the right
one), rather than one redis server and multiple redis namespaces (`-n`).