mirrorbits-geoupdate-all: Document a bug with this script
The only way to fix this is to improve the 'mirrorbits geoupdate'
command so that it can have a "automatic but conservative" mode.

Ie. the `geoupdate` command should have a way to prompt users only for
ASN, Country or Continent change, but apply Latitude and Longitude
changes automatically without asking.

Except that it's not clear that every user want that, maybe some would
like ASN changes to be applied automatically?

In the meantime, document the issue.