redis: Fix version check on startup
Since commit e61f9762, mirrorbits fails to start if ever the database is
not reachable (ie. redis is not running). Logs:

```
$ ./bin/mirrorbits daemon
 _______ __                        __     __ __
|   |   |__|.----.----.-----.----.|  |--.|__|  |_.-----.
|       |  ||   _|   _|  _  |   _||  _  ||  |   _|__ --|
|__|_|__|__||__| |__| |_____|__|  |_____||__|____|_____|  v0.5.1-113-g178f979

2025/03/27 02:12:21.207 UTC Unsupported Redis version, please upgrade to Redis >= 3.2.0
```

This is clearly a regression: by design, mirrorbits is able to run even
when the database is not reachable (in that case it redirects all the
requests to the fallback mirrors). And it worked before commit e61f9762.

Looking back at the code before e61f9762, we can see that mirrorbits
*could* fail to start like that, but in order to hit this failure mode,
the following conditions had to be met: 1) the connection with redis was
established and 2) the result of the command `INFO server` could be
parsed successfully and finally 3) the redis version was found to be
lower than the requirement.

With this commit, we restore this behavior.