redis: Remove dead code (CheckVersion and associated error)
The public function `CheckVersion()` is dead code since commit 14391551
(Nov 2018), when the redis version check was moved out of `main.go` and
into `database/redis.go`.

It was brought back to life recently in e61f9762, however as we've seen
in previous commit, it didn't cut it and we don't use it anymore, it's
dead again.

Let's cleanup and remove this function along with its associated error
`ErrRedisUpgradeRequired`.