redis: Make version parsing more robust
Before this fix, a version such as '6.0' would sort below '1.0.0' simply
because it doesn't have a patch number.

Arguably, not a bug, if you make the assumption that the redis version
ALWAYS have 3 components (major, minor and patch), no more no less, and
therefore we always compare versions that have 3 components.

Arguably, not a bug, because semver.org requires that those 3 components
are always present, and I suppose redis (and the various redis forks out
there) are careful to follow that.

At the same time, it feels weird to die on that hill. It's not very hard
to add zero values if missing, so that we support versions that have
either 0, 1, 2 or 3 components.

Note: after this change, an empty version "" becomes "0.0.0", which is
equivalent since it gets converted to a integer value of zero anyway.